+++ src/types/jest.ts
@@ -1,0 +1,9 @@
+declare global {
+	namespace jest {
+		interface Expect {
+			function(args: any[], expectedReturnValue: any): any;
+		}
+	}
+}
+
+export {};