+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1](../../tags/v2.1.1) - 2023-05-31
+### Added
+- New jest matcher to expect function
+### Changed
+- Update `@anmiles/google-api-wrapper` with breaking change
+- Use new jest matcher to expect function
+
 ## [2.1.0](../../tags/v2.1.0) - 2023-05-26
 ### Changed
 - Update `@anmiles/google-api-wrapper` in order to queue authentication in case of concurrent applications